requirements.txt
asgiref==3.2.9
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.7
idna==2.9
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
